# openapi.json
{
  "components": {
    "schemas": {
      "AgentMessage": {
        "properties": {
          "content": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "default": "user",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "AgentMessage",
        "type": "object"
      },
      "AgentQuestion": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AgentMessage"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "messages"
        ],
        "title": "AgentQuestion",
        "type": "object"
      },
...
      "Body_upload_examples_uc_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_examples_uc_upload_post",
        "type": "object"
      },
...
      "ExampleMessage": {
        "properties": {
          "text": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "ExampleMessage",
        "type": "object"
      },
      "ExampleRow": {
        "properties": {
          "data": {
            "maxLength": 65536,
            "title": "Data",
            "type": "string"
          },
          "id": {
            "maxLength": 255,
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "data"
        ],
        "title": "ExampleRow",
        "type": "object"
      },
      "ExampleTitle": {
        "properties": {
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "ExampleTitle",
        "type": "object"
      },
      "GenieQuestion": {
        "properties": {
          "content": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "GenieQuestion",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Databricks Apps FastAPI Starter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/examples/agent/ask": {
      "post": {
        "operationId": "agent_ask_examples_agent_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Agent Ask",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/agent/ask/stream": {
      "post": {
        "operationId": "agent_ask_stream_examples_agent_ask_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Agent Ask Stream",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/embed": {
      "post": {
        "operationId": "embed_examples_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleTitle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Embed",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/genie/{space_id}/ask": {
      "post": {
        "operationId": "genie_start_conversation_examples_genie__space_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenieQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Genie Start Conversation",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/genie/{space_id}/{conversation_id}/ask": {
      "post": {
        "operationId": "genie_follow_up_examples_genie__space_id___conversation_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenieQuestion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Genie Follow Up",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/job": {
      "post": {
        "operationId": "run_job_examples_job_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Params"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Run Job",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/pg": {
      "post": {
        "operationId": "pg_demo_examples_pg_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Pg Demo",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/serving": {
      "post": {
        "operationId": "serving_examples_serving_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ExampleRow"
                },
                "title": "Rows",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Serving",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/uc/download": {
      "get": {
        "operationId": "download_examples_uc_download_get",
        "parameters": [
          {
            "in": "query",
            "name": "relative_path",
            "required": true,
            "schema": {
              "title": "Relative Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Download",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/uc/upload": {
      "post": {
        "operationId": "upload_examples_uc_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "relative_path",
            "required": true,
            "schema": {
              "title": "Relative Path",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_examples_uc_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Upload",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/vector/query": {
      "post": {
        "operationId": "vector_query_examples_vector_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleTitle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Vector Query",
        "tags": [
          "examples"
        ]
      }
    },
    "/examples/vector/store": {
      "post": {
        "operationId": "vector_store_examples_vector_store_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleTitle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Vector Store",
        "tags": [
          "examples"
        ]
      }
    },
...
  }
}